requirements.txt
kivy==2.3.0
pyjnius==1.6.1
git+https://github.com/kivy/buildozer.git@master
